Fixed Gyro issue that was causing it to clamp to -360 to 360 instead of -180 to 180. This didn't affect teleop but caused issues for pathplanner as it's gyro readings are clamped to -180 180.
Added LED code and lighting priority system. Added LED conditions to swerveChassisSubsystem and several commands that rev up the shooter. Read the BlinkinLED file for a rundown on which lights take priority over others.

BlinkinLED is not a subsystem. The light is updated in robotPeriodic every update by calling the method through robotContainer